requirements.txt
anyio==3.6.2
asgiref==3.6.0
attrs==22.2.0
Babel==2.12.1
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.0
firebase-admin==6.0.1
Flask==2.2.2
Flask-Cors==3.0.10
google-api-core==2.11.0
google-api-python-client==2.74.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.9.1
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpcio==1.51.1
grpcio-status==1.51.1
h11==0.14.0
httpcore==0.16.3
httplib2==0.21.0
httpx==0.23.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.1
msgpack==1.0.4
nodeenv==1.7.0
prisma==0.8.0
proto-plus==1.22.2
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.4
PyJWT==2.6.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dotenv==0.21.0
requests==2.28.2
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
tomlkit==0.11.6
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.14
Werkzeug==2.2.2
